refactored, moved CmdLnRunner implementation and TestDatabaseSeeder classes to the 'test' package, configured application-test.properties to accommodate the H2 database for integration tests
