+++ pom.xml
@@ -79,6 +79,12 @@
 			<version>1.4.2.Final</version>
 			<scope>provided</scope>
 		</dependency>
+
+		<!-- h2database -->
+		<dependency>
+			<groupId>com.h2database</groupId>
+			<artifactId>h2</artifactId>
+		</dependency>
 	</dependencies>
 
 	<build>
