+++ src/main/resources/application-test.properties
@@ -1,15 +1,14 @@
 # server.port=8081
-# Connection url for the database
-spring.datasource.url=jdbc:mysql://localhost:3306/adeuxpas_integrationtest?serverTimezone=GMT
-spring.datasource.username=${MYSQL_USER}
-spring.datasource.password=${MYSQL_PASSWORD}
+# DataSource settings: H2
+spring.datasource.url=jdbc:h2:mem:testdb
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+spring.jpa.properties.hibernate.globally_quoted_identifiers=true
 
 # Show log for each sql query
 spring.jpa.show-sql=true
-
-# Hibernate ddl auto (create, create-drop, update): with "update" the database
-# schema will be automatically updated accordingly to java entities found in the project
-spring.jpa.hibernate.ddl-auto=create
 
 # seed users passwords
 first.pass = ${FIRST_PASS}
