+++ src/test/java/adeuxpas/back/testdatainit/CmdLinePopulateTestDatabase.java
@@ -1,4 +1,4 @@
-package adeuxpas.back.datainit;
+package adeuxpas.back.testdatainit;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.CommandLineRunner;
